refactor: move domain services to application layer
