+++ app/api/core/application/propertyCreatorService/PropertyCreatorService.ts
@@ -1,6 +1,6 @@
 import { Context, Property } from 'api/templates.v2/model/Property';
+import { PropertyFactory } from 'api/core/domain/template/PropertyFactory';
 import { AbstractPropertyCreatorService, CreateInput } from './AbstractPropertyCreatorService';
-import { PropertyFactory } from '../PropertyFactory';
 
 class PropertyCreatorService extends AbstractPropertyCreatorService {
   // eslint-disable-next-line class-methods-use-this
