+++ app/api/core/application/specs/NestedPropertyCreatorService.spec.ts
@@ -5,8 +5,8 @@ import { DefaultTransactionManager } from 'api/common.v2/database/data_source_de
 import { DefaultSettingsDataSource } from 'api/settings.v2/database/data_source_defaults';
 import { TestUtils } from 'api/common.v2/utils/Test';
 import { PropertyCreatorServiceStrategy } from '../propertyCreatorService/PropertyCreatorServiceStrategy';
-import { NestedPropertyNotAvailableError } from '../errors';
-import { NestedProperty } from '../NestedProperty';
+import { NestedPropertyNotAvailableError } from '../../domain/template/errors';
+import { NestedProperty } from '../../domain/template/NestedProperty';
 
 const createSut = () => {
   const transactionManager = DefaultTransactionManager();
