+++ app/api/core/application/specs/SelectPropertyCreatorService.spec.ts
@@ -4,7 +4,7 @@ import { DefaultTemplatesDataSource } from 'api/templates.v2/database/data_sourc
 import { DefaultTransactionManager } from 'api/common.v2/database/data_source_defaults';
 import { MongoThesauriDataSource } from 'api/core/infrastructure/mongodb/thesauri/MongoThesauriDS';
 import { SelectPropertyCreatorService } from '../propertyCreatorService/SelectPropertyCreatorService';
-import { SelectPropertyWithInvalidThesaurusError } from '../errors';
+import { SelectPropertyWithInvalidThesaurusError } from '../../domain/template/errors';
 
 describe('SelectPropertyCreatorService', () => {
   beforeAll(async () => {
